initialization clean up
